Moves util/interrupt into kubectl, since its only used there
