+++ pkg/kubectl/cmd/rollout/BUILD
@@ -26,7 +26,7 @@ go_library(
         "//pkg/kubectl/polymorphichelpers:go_default_library",
         "//pkg/kubectl/scheme:go_default_library",
         "//pkg/kubectl/util/i18n:go_default_library",
-        "//pkg/util/interrupt:go_default_library",
+        "//pkg/kubectl/util/interrupt:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/api/errors:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/api/meta:go_default_library",
         "//staging/src/k8s.io/apimachinery/pkg/apis/meta/v1:go_default_library",
