+++ pkg/kubectl/cmd/run/run.go
@@ -48,7 +48,7 @@ import (
 	"k8s.io/kubernetes/pkg/kubectl/polymorphichelpers"
 	"k8s.io/kubernetes/pkg/kubectl/scheme"
 	"k8s.io/kubernetes/pkg/kubectl/util/i18n"
-	"k8s.io/kubernetes/pkg/util/interrupt"
+	"k8s.io/kubernetes/pkg/kubectl/util/interrupt"
 	uexec "k8s.io/utils/exec"
 )
