+++ pkg/kubectl/util/BUILD
@@ -67,6 +67,7 @@ filegroup(
         ":package-srcs",
         "//pkg/kubectl/util/hash:all-srcs",
         "//pkg/kubectl/util/i18n:all-srcs",
+        "//pkg/kubectl/util/interrupt:all-srcs",
         "//pkg/kubectl/util/logs:all-srcs",
         "//pkg/kubectl/util/slice:all-srcs",
         "//pkg/kubectl/util/term:all-srcs",
