+++ pkg/kubectl/util/interrupt/BUILD
@@ -8,7 +8,7 @@ load(
 go_library(
     name = "go_default_library",
     srcs = ["interrupt.go"],
-    importpath = "k8s.io/kubernetes/pkg/util/interrupt",
+    importpath = "k8s.io/kubernetes/pkg/kubectl/util/interrupt",
 )
 
 filegroup(
