+++ pkg/kubectl/util/interrupt/interrupt.go
