+++ pkg/util/BUILD
@@ -27,7 +27,6 @@ filegroup(
         "//pkg/util/goroutinemap:all-srcs",
         "//pkg/util/hash:all-srcs",
         "//pkg/util/initsystem:all-srcs",
-        "//pkg/util/interrupt:all-srcs",
         "//pkg/util/io:all-srcs",
         "//pkg/util/ipconfig:all-srcs",
         "//pkg/util/ipset:all-srcs",
